Remove original input module classes
