+++ cellprofiler_core/constants/image.py
@@ -110,3 +110,20 @@ CT_COLOR = "Color"
 CT_MASK = "Mask"
 CT_OBJECTS = "Objects"
 CT_FUNCTION = "Function"
+
+MD_PLANAR = "Planar"
+MD_SIZE_S = "SizeS"
+MD_SIZE_C = "SizeC"
+MD_SIZE_Z = "SizeZ"
+MD_SIZE_T = "SizeT"
+MD_SIZE_X = "SizeX"
+MD_SIZE_Y = "SizeY"
+
+MD_CHANNEL_NAME = "ChannelName"
+MD_COLOR_FORMAT = "ColorFormat"
+MD_MONOCHROME = "monochrome"
+MD_RGB = "RGB"
+
+MD_C = "C"
+MD_Z = "Z"
+MD_T = "T"
